test(BA-7010): move the scoped upsert tests next to the other write ops
upsert_scoped is a method on RBACWriteOps like create_scope and add_entity_members,
so its tests belong in the ops provider test file rather than in one of their own.
